Move typing helpers to utils package
