+++ docs/api/scikit_build_core.utils.rst
@@ -5,3 +5,14 @@ scikit\_build\_core.utils package
    :members:
    :show-inheritance:
    :undoc-members:
+
+Submodules
+----------
+
+scikit\_build\_core.utils.typing module
+---------------------------------------
+
+.. automodule:: scikit_build_core.utils.typing
+   :members:
+   :show-inheritance:
+   :undoc-members:
