+++ src/scikit_build_core/settings/sources.py
@@ -49,26 +49,21 @@ from __future__ import annotations
 
 import dataclasses
 import os
-import sys
 import typing
-from typing import Any, Literal, Protocol, TypeVar, Union
+from typing import Any, Literal, Protocol, TypeVar
 
 from .._compat.builtins import ExceptionGroup
-from .._compat.typing import Annotated, get_args, get_origin
+from .._compat.typing import get_args
+from ..utils.typing import (
+    _get_inner_type,
+    _get_target_raw_type,
+    _is_union_type,
+    _process_annotated,
+    _process_union,
+)
 
 if typing.TYPE_CHECKING:
     from collections.abc import Generator, Iterator, Mapping, Sequence
-
-# Runtime check for PEP 604 union syntax (A | B) support
-# types.UnionType only exists in Python 3.10+
-if sys.version_info >= (3, 10):
-    from types import NoneType
-    from types import UnionType as _UnionType
-else:
-    NoneType = type(None)
-
-    class _UnionType:
-        pass
 
 
 T = TypeVar("T")
@@ -105,71 +100,6 @@ def _dig_fields(opt: Any, /, *names: str) -> Any:
             raise KeyError(msg)
         (opt,) = types
     return opt
-
-
-def _process_union(target: Any, /) -> Any:
-    """
-    Filters None out of Unions. If a Union only has one item, return that item.
-    """
-
-    origin = get_origin(target)
-
-    if _is_union_type(origin):
-        non_none_args = [a for a in get_args(target) if a is not NoneType]
-        if len(non_none_args) == 1:
-            return non_none_args[0]
-        return Union[tuple(non_none_args)]
-
-    return target
-
-
-def _process_annotated(target: type[Any], /) -> tuple[Any, tuple[Any, ...]]:
-    """
-    Splits annotated into raw type and annotations. If not annotated, the annotations will be empty.
-    """
-
-    origin = get_origin(target)
-    if origin is Annotated:
-        return get_args(target)[0], get_args(target)[1:]
-
-    return target, ()
-
-
-def _get_target_raw_type(target: type[Any] | Any, /) -> Any:
-    """
-    Takes a type like ``Optional[str]`` and returns str, or ``Optional[Dict[str,
-    int]]`` and returns dict. Returns Union for a Union with more than one
-    non-none type. Literal is also a valid return. Works through Annotated.
-    """
-
-    target, _ = _process_annotated(target)
-    target = _process_union(target)
-    origin = get_origin(target)
-    return origin or target
-
-
-def _is_union_type(raw_target: Any) -> bool:
-    """
-    Check if raw_target is a Union type (either ``typing.Union`` or ``types.UnionType``).
-    Handles both ``typing.Union[A, B]`` and PEP 604 syntax (``A | B``).
-    """
-    return raw_target is Union or raw_target is _UnionType
-
-
-def _get_inner_type(_target: type[Any], /) -> type[Any]:
-    """
-    Takes a types like ``List[str]`` and returns str,
-    or ``Dict[str, int]`` and returns int.
-    """
-
-    raw_target = _get_target_raw_type(_target)
-    target = _process_union(_target)
-    if raw_target is list:
-        return get_args(target)[0]  # type: ignore[no-any-return]
-    if raw_target is dict:
-        return get_args(target)[1]  # type: ignore[no-any-return]
-    msg = f"Expected a list or dict, got {target!r}"
-    raise AssertionError(msg)
 
 
 def _nested_dataclass_to_names(
